requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.4
django-cors-headers==2.4.0
django-rest-knox==3.4.0
django-sendgrid-v5==0.7.0
django-webpack-loader==0.6.0
djangorestframework==3.9.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2==2.7.6.1
pycparser==2.19
pyOpenSSL==18.0.0
python-http-client==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
sendgrid==5.6.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
